changing the like function in the model
this commit changes the null.stan file so that the like function will use a vector of Qdiff. This speeds up the sampling to about a third then what it was.
Note that the 'selected_offer' variable in the data is now coded 0 for the first offer and 1 for the second.